pyproject.toml
[project]
name = "dnd-combat-sim"
version = "0.1.0"
description = "D&D 2024 combat simulator for build comparison"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.ruff]
line-length = 100
